+++ api/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.21.1] - 2021-11-19
+
+### Changed
+
+- Updated `ffi_common` to 0.7.0.
+
+### Fixed
+
+- gh-35: Use a separate rule for parsing annotations--one that permits more characters.
+
 ## [0.21.0] - 2021-10-20
 
 ### Added